feature: gltf: gltf object, initialization from scenefiles, cleanup
