+++ clovers-cli/Cargo.toml
@@ -10,7 +10,7 @@ path = "src/main.rs"
 
 [dependencies]
 # Internal
-clovers = { path = "../clovers", features = ["serde-derive", "stl", "traces"], default-features = false }
+clovers = { path = "../clovers", features = ["serde-derive", "stl", "traces", "gl_tf"], default-features = false }
 
 # External
 clap = { version = "4.0.8", features = ["std", "derive"], default-features = false  }
@@ -25,4 +25,3 @@ serde_json = { version = "1.0", features = ["alloc"], default-features = false }
 time = { version = "0.3.14", default-features = false }
 tracing = "0.1.36"
 tracing-subscriber = { version = "0.3.15", features = ["time"] }
-gltf = { version = "1.0.0", features = ["import"] }
